requirements.txt
aiofiles
aiohttp
asyncio
beautifulsoup4
dnspython
ffmpeg-python
gitpython
hachoir
heroku3
lyricsgenius
motor
numpy
pillow
psutil
py-tgcalls==0.9.7
pykeyboard
pyrofork
# pyrogram==2.0.106
python-dotenv
pyyaml
requests
speedtest-cli
spotipy
tgcrypto
unidecode
yt-dlp
youtube-search
youtube-search-python
